requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==35.0.0
dateparser==1.1.0
frozenlist==1.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.2.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==21.0.0
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.2
requests==2.26.0
rsa==4.7.2
service-identity==21.1.0
six==1.16.0
Telethon==1.23.0
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.2
zope.interface==5.4.0
